AUT-804: Add test helper to convert CSV to User Profile and User Credentials - Add test helper to convert CSV to User Profile and User Credentials - This is needed for both create and delete lambda tests
